Renaming test_util to testutil
