+++ go/pkg/filemetadata/BUILD.bazel
@@ -30,7 +30,7 @@ go_test(
     deps = [
         "//go/pkg/cache:go_default_library",
         "//go/pkg/digest:go_default_library",
-        "//go/pkg/test_util:go_default_library",
+        "//go/pkg/testutil:go_default_library",
         "@com_github_google_go_cmp//cmp:go_default_library",
         "@com_github_google_go_cmp//cmp/cmpopts:go_default_library",
     ],
