+++ go/pkg/filemetadata/cache_posix_test.go
@@ -7,13 +7,13 @@ import (
 	"os"
 	"testing"
 
-	"github.com/bazelbuild/remote-apis-sdks/go/pkg/test_util"
+	"github.com/bazelbuild/remote-apis-sdks/go/pkg/testutil"
 	"github.com/google/go-cmp/cmp"
 )
 
 func TestExecutableCacheLoad(t *testing.T) {
 	c := NewSingleFlightCache()
-	filename, err := test_util.CreateFile(t, true, "")
+	filename, err := testutil.CreateFile(t, true, "")
 	if err != nil {
 		t.Fatalf("Failed to create tmp file for testing digests: %v", err)
 	}
