+++ go/pkg/reader/BUILD.bazel
@@ -12,7 +12,7 @@ go_test(
     srcs = ["reader_test.go"],
     embed = [":go_default_library"],
     deps = [
-        "//go/pkg/test_util:go_default_library",
+        "//go/pkg/testutil:go_default_library",
         "@com_github_google_go_cmp//cmp:go_default_library",
     ],
 )
