+++ go/pkg/testutil/BUILD.bazel
@@ -2,7 +2,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = ["test_util.go"],
-    importpath = "github.com/bazelbuild/remote-apis-sdks/go/pkg/test_util",
+    srcs = ["testutil.go"],
+    importpath = "github.com/bazelbuild/remote-apis-sdks/go/pkg/testutil",
     visibility = ["//visibility:public"],
 )
