+++ go/pkg/testutil/testutil.go
@@ -1,4 +1,4 @@
-package test_util
+package testutil
 
 import (
 	"io/ioutil"
